backend/requirements.txt
Flask==2.3.3
opencensus-ext-azure==1.1.9